test: move tests related to sign to sign.e2e.ts
