+++ e2e/harmony/sign.e2e.ts
@@ -1,6 +1,13 @@
+import fs from 'fs';
+import path from 'path';
 import chai, { expect } from 'chai';
 import { Extensions } from '@teambit/legacy.constants';
-import { Helper } from '@teambit/legacy.e2e-helper';
+import { Helper, NpmCiRegistry, supportNpmCiRegistryTesting } from '@teambit/legacy.e2e-helper';
+import { DEPS_GRAPH } from '@teambit/harmony.modules.feature-toggle';
+import { type BitLockfileFile } from '@teambit/pnpm';
+import { generateRandomStr } from '@teambit/toolbox.string.random';
+import stripAnsi from 'strip-ansi';
+import { addDistTag } from '@pnpm/registry-mock';
 
 chai.use(require('chai-fs'));
 
@@ -260,6 +267,239 @@ describe('sign command', function () {
     it('should save updated versions on the remotes', () => {
       const comp1 = helper.command.catComponent(`${helper.scopes.remote}/comp1@latest`, helper.scopes.remotePath);
       expect(comp1.buildStatus).to.equal('succeed');
+    });
+  });
+  (supportNpmCiRegistryTesting ? describe : describe.skip)('dependencies graph data', function () {
+    this.timeout(0);
+    let npmCiRegistry: NpmCiRegistry;
+    before(() => {
+      helper = new Helper();
+      helper.command.setFeatures(DEPS_GRAPH);
+    });
+    after(() => {
+      helper.scopeHelper.destroy();
+      helper.command.resetFeatures();
+    });
+    describe('single component', () => {
+      before(async () => {
+        helper.scopeHelper.setWorkspaceWithRemoteScope();
+        npmCiRegistry = new NpmCiRegistry(helper);
+        npmCiRegistry.configureCiInPackageJsonHarmony();
+        await npmCiRegistry.init();
+        helper.command.setConfig('registry', npmCiRegistry.getRegistryUrl());
+        helper.fixtures.populateComponents(1);
+        helper.fs.outputFile(`comp1/index.js`, `const isEven = require("is-even"); require('@pnpm.e2e/pkg-with-1-dep')`);
+        helper.fs.outputFile(
+          `comp1/index.spec.js`,
+          `const isOdd = require("is-odd"); test('test', () => { expect(1).toEqual(1); })`
+        );
+        await addDistTag({ package: '@pnpm.e2e/pkg-with-1-dep', version: '100.0.0', distTag: 'latest' });
+        await addDistTag({ package: '@pnpm.e2e/dep-of-pkg-with-1-dep', version: '100.0.0', distTag: 'latest' });
+        helper.workspaceJsonc.addKeyValToDependencyResolver('policy', {
+          dependencies: {
+            '@pnpm.e2e/pkg-with-1-dep': '^100.0.0',
+          },
+        });
+        helper.command.install('is-even@1.0.0 is-odd@1.0.0');
+        helper.command.snapAllComponentsWithoutBuild('--skip-tests');
+        await addDistTag({ package: '@pnpm.e2e/pkg-with-1-dep', version: '100.1.0', distTag: 'latest' });
+        await addDistTag({ package: '@pnpm.e2e/dep-of-pkg-with-1-dep', version: '100.1.0', distTag: 'latest' });
+      });
+      after(() => {
+        npmCiRegistry.destroy();
+        helper.command.delConfig('registry');
+      });
+      it('should save dependencies graph to the model', () => {
+        const versionObj = helper.command.catComponent('comp1@latest');
+        const depsGraph = JSON.parse(helper.command.catObject(versionObj.dependenciesGraphRef));
+        const directDeps = depsGraph.edges.find((edge) => edge.id === '.')?.neighbours;
+        expect(directDeps).deep.include({
+          name: 'is-even',
+          specifier: '1.0.0',
+          id: 'is-even@1.0.0',
+          lifecycle: 'runtime',
+          optional: false,
+        });
+        expect(directDeps).deep.include({
+          name: 'is-odd',
+          specifier: '1.0.0',
+          id: 'is-odd@1.0.0',
+          lifecycle: 'dev',
+          optional: false,
+        });
+      });
+      describe('sign component and use dependency graph to generate a lockfile', () => {
+        let signOutput: string;
+        let signRemote;
+        before(async () => {
+          helper.command.export();
+          signRemote = helper.scopeHelper.getNewBareScope('-remote-sign');
+          helper.scopeHelper.addRemoteScope(helper.scopes.remotePath, signRemote.scopePath);
+          const { head } = helper.command.catComponent(`${helper.scopes.remote}/comp1`);
+          signOutput = helper.command.sign(
+            [`${helper.scopes.remote}/comp1@${head}`],
+            '--push --log',
+            signRemote.scopePath
+          );
+        });
+        it('should sign successfully', () => {
+          expect(signOutput).to.include('the following 1 component(s) were signed with build-status "succeed"');
+        });
+        it('should save dependencies graph to the model', () => {
+          const versionObj = helper.command.catComponent('comp1@latest');
+          const depsGraph = JSON.parse(helper.command.catObject(versionObj.dependenciesGraphRef));
+          const directDeps = depsGraph.edges.find((edge) => edge.id === '.')?.neighbours;
+          expect(directDeps).deep.include({
+            name: 'is-even',
+            specifier: '1.0.0',
+            id: 'is-even@1.0.0',
+            lifecycle: 'runtime',
+            optional: false,
+          });
+          expect(directDeps).deep.include({
+            name: 'is-odd',
+            specifier: '1.0.0',
+            id: 'is-odd@1.0.0',
+            lifecycle: 'dev',
+            optional: false,
+          });
+        });
+      });
+      describe('imported component uses dependency graph to generate a lockfile', () => {
+        before(async () => {
+          helper.scopeHelper.reInitWorkspace();
+          helper.scopeHelper.addRemoteScope();
+          helper.command.import(`${helper.scopes.remote}/comp1@latest`);
+        });
+        it('should generate a lockfile', () => {
+          expect(fs.readFileSync(path.join(helper.scopes.localPath, 'pnpm-lock.yaml'), 'utf8')).to.have.string(
+            'restoredFromModel: true'
+          );
+        });
+      });
+    });
+    describe('single component and sign writes the dependency graph', () => {
+      before(async () => {
+        helper.scopeHelper.setWorkspaceWithRemoteScope();
+        npmCiRegistry = new NpmCiRegistry(helper);
+        npmCiRegistry.configureCiInPackageJsonHarmony();
+        await npmCiRegistry.init();
+        helper.command.setConfig('registry', npmCiRegistry.getRegistryUrl());
+        helper.fixtures.populateComponents(1);
+        helper.fs.outputFile(`comp1/index.js`, `const React = require("react"); require('@pnpm.e2e/pkg-with-1-dep')`);
+        helper.fs.outputFile(
+          `comp1/index.spec.js`,
+          `const isOdd = require("is-odd"); test('test', () => { expect(1).toEqual(1); })`
+        );
+        await addDistTag({ package: '@pnpm.e2e/pkg-with-1-dep', version: '100.0.0', distTag: 'latest' });
+        await addDistTag({ package: '@pnpm.e2e/dep-of-pkg-with-1-dep', version: '100.0.0', distTag: 'latest' });
+        helper.workspaceJsonc.addKeyValToDependencyResolver('policy', {
+          dependencies: {
+            '@pnpm.e2e/pkg-with-1-dep': '^100.0.0',
+          },
+        });
+        helper.command.install('react@18.3.1 is-odd@1.0.0');
+        fs.unlinkSync(path.join(helper.fixtures.scopes.localPath, 'pnpm-lock.yaml'));
+        helper.command.snapAllComponentsWithoutBuild('--skip-tests');
+        await addDistTag({ package: '@pnpm.e2e/pkg-with-1-dep', version: '100.1.0', distTag: 'latest' });
+        await addDistTag({ package: '@pnpm.e2e/dep-of-pkg-with-1-dep', version: '100.1.0', distTag: 'latest' });
+      });
+      after(() => {
+        npmCiRegistry.destroy();
+        helper.command.delConfig('registry');
+      });
+      it('should not save dependencies graph to the model', () => {
+        const versionObj = helper.command.catComponent('comp1@latest');
+        expect(versionObj.dependenciesGraphRef).to.be.undefined;
+      });
+      describe('sign component and use dependency graph to generate a lockfile', () => {
+        let signOutput: string;
+        let lockfile: BitLockfileFile;
+        let signRemote;
+        before(async () => {
+          helper.command.export();
+          signRemote = helper.scopeHelper.getNewBareScope('-remote-sign');
+          helper.scopeHelper.addRemoteScope(helper.scopes.remotePath, signRemote.scopePath);
+          const { head } = helper.command.catComponent(`${helper.scopes.remote}/comp1`);
+          signOutput = helper.command.sign(
+            [`${helper.scopes.remote}/comp1@${head}`],
+            '--push --log',
+            signRemote.scopePath
+          );
+        });
+        it('should sign successfully', () => {
+          expect(signOutput).to.include('the following 1 component(s) were signed with build-status "succeed"');
+        });
+        it('should generate a lockfile', () => {
+          const capsulesDir = signOutput.match(/running installation in root dir (\/[^\s]+)/)?.[1];
+          expect(capsulesDir).to.be.a('string');
+          lockfile = yaml.load(fs.readFileSync(path.join(stripAnsi(capsulesDir!), 'pnpm-lock.yaml'), 'utf8'));
+          expect(lockfile.bit).to.eql({
+            depsRequiringBuild: [],
+          });
+        });
+      });
+      describe('imported component uses dependency graph to generate a lockfile', () => {
+        before(async () => {
+          helper.scopeHelper.reInitWorkspace();
+          helper.scopeHelper.addRemoteScope();
+          helper.command.import(`${helper.scopes.remote}/comp1@latest`);
+        });
+        it('should generate a lockfile', () => {
+          expect(fs.readFileSync(path.join(helper.scopes.localPath, 'pnpm-lock.yaml'), 'utf8')).to.have.string(
+            'restoredFromModel: true'
+          );
+        });
+        it('should save dependencies graph to the model', () => {
+          const versionObj = helper.command.catComponent('comp1@latest');
+          const depsGraph = JSON.parse(helper.command.catObject(versionObj.dependenciesGraphRef));
+          expect(depsGraph).to.not.be.undefined;
+        });
+      });
+    });
+    describe('two components exported then one imported', function () {
+      let randomStr: string;
+      before(async () => {
+        randomStr = generateRandomStr(4); // to avoid publishing the same package every time the test is running
+        const name = `@ci/${randomStr}.{name}`;
+        helper.scopeHelper.setWorkspaceWithRemoteScope();
+        npmCiRegistry = new NpmCiRegistry(helper);
+        npmCiRegistry.configureCustomNameInPackageJsonHarmony(name);
+        await npmCiRegistry.init();
+        helper.command.setConfig('registry', npmCiRegistry.getRegistryUrl());
+
+        helper.fixtures.populateComponents(2);
+        helper.command.install('--add-missing-deps');
+        helper.command.tagAllWithoutBuild('--skip-tests');
+        helper.command.export();
+        const signRemote = helper.scopeHelper.getNewBareScope('-remote-sign');
+        helper.scopeHelper.addRemoteScope(helper.scopes.remotePath, signRemote.scopePath);
+        helper.command.sign(
+          [`${helper.scopes.remote}/comp2@0.0.1`],
+          '',
+          signRemote.scopePath
+        );
+        helper.command.export();
+        helper.command.sign(
+          [`${helper.scopes.remote}/comp1@0.0.1`],
+          '',
+          signRemote.scopePath
+        );
+        helper.command.export();
+      });
+      after(() => {
+        npmCiRegistry.destroy();
+        helper.command.delConfig('registry');
+        helper.scopeHelper.destroy();
+      });
+      it('should generate a lockfile', () => {
+        helper.scopeHelper.reInitWorkspace();
+        helper.scopeHelper.addRemoteScope();
+        helper.command.import(`${helper.scopes.remote}/comp1@latest`);
+        expect(helper.fs.readFile('pnpm-lock.yaml')).to.have.string(
+          'restoredFromModel: true'
+        );
+      });
     });
   });
 });
